Centralise FacilityModel
