+++ packages/meditrak-server/src/database/models/Facility.js
@@ -1,36 +1,11 @@
 /**
- * Tupaia MediTrak
- * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
- **/
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
 
-import { DatabaseModel, DatabaseType, TYPES } from '@tupaia/database';
+import { FacilityModel as CommonFacilityModel } from '@tupaia/database';
 
-class FacilityType extends DatabaseType {
-  static databaseType = TYPES.FACILITY;
-
-  // Exposed for access policy creation.
-  get organisationUnitCode() {
-    return this.code;
-  }
-
-  async geographicalArea() {
-    return this.otherModels.geographicalArea.findById(this.geographical_area_id);
-  }
-
-  async country() {
-    return this.otherModels.country.findById(this.country_id);
-  }
-
-  async entity() {
-    return this.otherModels.entity.findOne({ code: this.code });
-  }
-}
-
-export class FacilityModel extends DatabaseModel {
-  get DatabaseTypeClass() {
-    return FacilityType;
-  }
-
+export class FacilityModel extends CommonFacilityModel {
   meditrakConfig = {
     minAppVersion: '0.0.1',
   };
