+++ packages/web-config-server/src/models/index.js
@@ -5,8 +5,6 @@
 
 import { ProjectModel } from './Project';
 import { DashboardGroupModel } from './DashboardGroup';
-export { UserSession } from './UserSession';
-export { Facility } from './Facility';
 export { EntityRelation } from './EntityRelation';
 export { Entity } from './Entity';
 export { AncestorDescendantRelation } from './AncestorDescendantRelation';
